I do not need to set ExecStop
https://unix.stackexchange.com/questions/426042/systemd-define-a-service-without-execstop-and-be-able-to-stop-it-without-fail

When systemd tries to stop a process without ExecStop= it sends a KillSignal= which defaults to SIGTERM (terminate signal), waits for a response from the process, and if it fails to stop within a deadline, systemd sends a SIGKILL which simply forcefully ends the process. systemd's behaviour can be modified with KillSignal= (see man systemd.kill https://www.freedesktop.org/software/systemd/man/systemd.kill.html).